requirements.txt
torch==1.8.1
torchvision==0.9.1
matplotlib==3.4.1
Pillow==8.1.0
numpy==1.20.1
pandas==1.2.2
tqdm==4.56.2
